server/pyproject.toml
[project]
name = "voice-dictation"
version = "0.1.0"
description = "Customizable AI-powered voice dictation tool"
license = "AGPL-3.0"
readme = "../README.md"
requires-python = ">=3.13"
dependencies = [
    "pipecat-ai[assemblyai,anthropic,cartesia,cerebras,deepgram,google,groq,openai,silero]>=0.0.96",
    "pydantic-settings>=2.12.0",
    "loguru>=0.7.3",
    "typer>=0.20.0",
    "fastapi>=0.123.9",
    "uvicorn>=0.38.0",
]

[project.scripts]
dictation-server = "dictation_server:app"

[dependency-groups]
dev = ["ruff>=0.14.3", "ty<=0.0.1a25"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "config", "services", "utils", "processors"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (includes unused imports/variables F401, F841)
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "PTH", # flake8-use-pathlib
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line too long, handled by formatter
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "RET504", # unnecessary return
    "ANN401", # dynamically typed expressions (Any) - needed for **kwargs passthrough
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.rules]
division-by-zero = "error"
deprecated = "error"
unused-ignore-comment = "error"
possibly-unresolved-reference = "error"
